+++ api_yamdb/api/permissions.py
@@ -13,3 +13,15 @@ class CheckUser(permissions.BasePermission):
 
     def has_object_permission(self, request, view, obj):
         return request.user == obj
+
+
+class IsAdminModeratorOwnerOrReadOnly(permissions.BasePermission):
+    def has_object_permission(self, request, view, obj):
+        return (request.method in permissions.SAFE_METHODS
+                or request.user.is_admin
+                or request.user.is_moderator
+                or obj.author == request.user)
+
+    def has_permission(self, request, view):
+        return (request.method in permissions.SAFE_METHODS
+                or request.user.is_authenticated)
