--- api_yamdb/reviews/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from .models import Review, Comment
-
-admin.site.register(Review)
-admin.site.register(Comment)
